Migrate Nimbus to Device/Driver/CapabilityBackend architecture
Split the monolithic NimbusBackend (~2300 lines) into the new PLR architecture:

- NimbusDriver(Driver): TCP I/O, protocol init, device-level ops (park, door lock)
- NimbusPIPBackend(PIPBackend): protocol translation for liquid handling
- Nimbus(Device): wires driver + PIP capability frontend
- NimbusChatterboxDriver: no-op driver for testing without hardware
- commands.py: all HamiltonCommand subclasses extracted to shared module

The legacy NimbusBackend is rewritten as a thin wrapper that delegates to
the new classes, preserving backwards compatibility. All 66 existing tests pass.
